Startknop toegevoegd die naar volgend scherm lijdt
