+++ Legpuzzel_ver1_Meindert/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:Legpuzzel_ver1_Meindert"
-             StartupUri="MainWindow.xaml">
+             StartupUri="StartScherm.xaml">
     <Application.Resources>
          
     </Application.Resources>
