+++ Legpuzzel_ver1_Meindert/Legpuzzel_ver1_Meindert.csproj.user
@@ -12,7 +12,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Page Update="MainWindow.xaml">
+    <Page Update="StartScherm.xaml">
       <SubType>Designer</SubType>
     </Page>
     <Page Update="NaStartscherm.xaml">
