--- Legpuzzel_ver1_Meindert/MainWindow.xaml
@@ -1,132 +1,0 @@
-﻿<Window x:Class="Legpuzzel_ver1_Meindert.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Legpuzzel_ver1_Meindert"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-    <Window.Resources>
-        <ImageBrush x:Key="MusicOff" ImageSource="\Pictures\MusicOff.png" />
-        <ImageBrush x:Key="MusicOn"  ImageSource="\Pictures\MusicOn.png" />
-        <ImageBrush x:Key="MusicOffHover" ImageSource="\Pictures\MusicOffHover.png" />
-        <ImageBrush x:Key="MusicOnHover"  ImageSource="\Pictures\MusicOnHover.png" />
-        <ImageBrush x:Key="OffButton" ImageSource="\Pictures\OffButton.png" />
-        <ImageBrush x:Key="OffButtonHover"  ImageSource="\Pictures\OffButtonHover.png" />
-
-
-
-        <Style x:Key="NoBugMusicOffStyle" TargetType="Button">
-            <Setter Property="Background" Value="{StaticResource MusicOff}" />
-            <Setter Property="BorderBrush" Value="Transparent" />
-            <Setter Property="Foreground" Value="Black" />
-            <Setter Property="Cursor" Value="Arrow" />
-            <Setter Property="Template">
-                
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}"
-                        BorderBrush="{TemplateBinding BorderBrush}"
-                        BorderThickness="{TemplateBinding BorderThickness}">
-                            <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center" />
-                        </Border>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="{StaticResource MusicOffHover}" />
-                </Trigger>
-            </Style.Triggers>
-        </Style>
-
-        <Style x:Key="NoBugMusicOnStyle" TargetType="Button">
-            <Setter Property="Background" Value="{StaticResource MusicOn}" />
-            <Setter Property="BorderBrush" Value="Transparent" />
-            <Setter Property="Foreground" Value="Black" />
-            <Setter Property="Cursor" Value="Arrow" />
-            <Setter Property="Template">
-
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}"
-                BorderBrush="{TemplateBinding BorderBrush}"
-                BorderThickness="{TemplateBinding BorderThickness}">
-                            <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center" />
-                        </Border>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="{StaticResource MusicOnHover}" />
-                </Trigger>
-            </Style.Triggers>
-        </Style>
-
-        <Style x:Key="NoBugOffButtonStyle" TargetType="Button">
-            <Setter Property="Background" Value="{StaticResource OffButton}" />
-            <Setter Property="BorderBrush" Value="Transparent" />
-            <Setter Property="Foreground" Value="Black" />
-            <Setter Property="Cursor" Value="Arrow" />
-            <Setter Property="Template">
-
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Border Background="{TemplateBinding Background}"
-        BorderBrush="{TemplateBinding BorderBrush}"
-        BorderThickness="{TemplateBinding BorderThickness}">
-                            <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center" />
-                        </Border>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-            <Style.Triggers>
-                <Trigger Property="IsMouseOver" Value="True">
-                    <Setter Property="Background" Value="{StaticResource OffButtonHover}" />
-                </Trigger>
-            </Style.Triggers>
-        </Style>
-        <Style x:Key="Buttons" TargetType="Button">
-            <Setter Property="Background" Value="#FF6666"/>
-            <Setter Property="FontSize" Value="40"/>
-            <Setter Property="FontWeight" Value="Bold"/>
-            <Setter Property="Foreground" Value="#F5F5F5"/>
-            <Setter Property="BorderThickness" Value="0"/>
-            <Setter Property="BorderBrush" Value="#FF6666"/>
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Grid>
-                            <Border Background="{TemplateBinding Background}"
-                                    BorderBrush="{TemplateBinding BorderBrush}"
-                                    BorderThickness="{TemplateBinding BorderThickness}"
-                                    CornerRadius="30">
-                                <ContentPresenter HorizontalAlignment="Center" VerticalAlignment="Center"/>
-                            </Border>
-                        </Grid>
-                        <ControlTemplate.Triggers>
-                            <Trigger Property="IsMouseOver" Value="True">
-                                <Setter Property="Background" Value="#FF9966"/>
-                                <Setter Property="BorderBrush" Value="#FF9966"/>
-                            </Trigger>
-                        </ControlTemplate.Triggers>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-    </Window.Resources>
-    
-
-    <Grid>
-        <Rectangle Fill="LightPink"/>
-        <Image x:Name="LOGO" HorizontalAlignment="Center" Height="100"  VerticalAlignment="Center" Width="100" Source="/Pictures/logo_carrot_gems.png" />
-        <Button Name="MuziekKnop" HorizontalAlignment="Left" VerticalAlignment="Top" Height="100" Width="100" Click="MuziekKnop_Click"  BorderBrush="Transparent" Style="{StaticResource NoBugMusicOffStyle}"/>
-
-        <MediaElement x:Name="mediaElement" Source="muziek\De_muziek2.mp3" LoadedBehavior="Manual"  MediaEnded="MediaElement_MediaEnded"/>
-        <Button x:Name="ExitButton" HorizontalAlignment="Right" VerticalAlignment="Top" Height="100" Width="100" Click="ExitButton_Click" BorderBrush="Transparent" Style="{StaticResource NoBugOffButtonStyle}"/>
-        <Button x:Name="StartButton" Content="Start" HorizontalAlignment="Center" VerticalAlignment="Center" Height="60" Width="170" Margin="0 250 0 0" Click="StartButton_Click" Style="{StaticResource Buttons}"/>
-
-
-    </Grid>
-</Window>
